Deprecate `Init(op)` (#47)
Using `foldl(op, xf, collection; init = Init)` in Transducers.jl has
been very useful.  However, it requires a special handling of `Init`
because it is not in the "initializer" type hierarchy:

> ```julia
> initialize(::typeof(Init), op) = check_init(Init(op), Init, op)
> initialize(f::InitOf, op) = check_init(f(op), f, op)
> ```
> --- https://github.com/JuliaFolds/Transducers.jl/blob/e5bc17538f16a0cc2d89e128bc509b8b1efa6bfc/src/core.jl#L873-L874

A (failed) solution attempted was to move the initializer API to
InitialValues.jl:

https://github.com/JuliaFolds/InitialValues.jl/pull/35

However, it didn't work well because the initializer is tightly
coupled with `Transducers.start` API.  It looks like moving this to
Transducers.jl (or "FoldsBase.jl") would be better:

* https://github.com/JuliaFolds/InitialValues.jl/pull/37: revert #35
* https://github.com/JuliaFolds/Transducers.jl/pull/289#issuecomment-633291202:
Explains the rationale to revert #35.

In order to treat `Init` in a unified "initializer" framework while
keeping InitialValues.jl self-contained, this PR starts the process of
removing `Init` from InitialValues.jl by deprecating it.  The new
interface is simply `InitialValue(op)`.  This is nice also because
`InitialValue(op) isa InitialValue` holds.

See also:
https://github.com/JuliaFolds/Transducers.jl/pull/341
(Refactoring initializer API; define `Init` in Transducers.jl)